refactor(BlackjackStatistics를): BlackjackStatistics를 result 패키지로 이동
