+++ src/main/java/domain/game/result/PlayerProfit.java
